wip move common logic to content workspace context base
