+++ src/packages/core/content/workspace/index.ts
@@ -1,3 +1,4 @@
-export type * from './content-workspace-context.interface.js';
+export * from './content-detail-workspace-base.js';
 export * from './content-workspace.context-token.js';
 export * from './views/edit/index.js';
+export type * from './content-workspace-context.interface.js';
